requirements.txt
albumentations==0.2.3
audioread==2.1.8
certifi==2020.6.20
cycler==0.10.0
decorator==4.4.0
imageio==2.5.0
imgaug==0.2.6
joblib==0.13.2
kafka-python==2.0.1
kiwisolver==1.1.0
librosa==0.6.3
llvmlite==0.29.0
matplotlib==3.1.0
networkx==2.3
numba==0.44.0
numpy==1.16.4
opencv-python-headless==4.1.0.25
oyaml==0.9
pandas==0.24.2
Pillow==6.0.0
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1.1
resampy==0.2.1
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
tqdm==4.32.1
